# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bsllmner-viewer API",
    "description": "Queries, aggregations, and exports over ontology-mapped BioSample annotations. The conventions follow the DDBJ Search API.\n\nStart with [/llms.txt](/llms.txt): condition examples and recipes for common tasks. [/llms-full.txt](/llms-full.txt) has the rules that span operations: the condition language, the descendants of terms, counting units and populations, self-exclusion, default elements, expected counts, export formats, errors, limits, and caching with `ETag` and `If-None-Match`. A description in this document names the heading of the rule that it relies on.",
    "contact": {
      "name": "BioData Science Initiative",
      "url": "https://github.com/dbcls/bsllmner-viewer"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/terms": {
      "get": {
        "tags": [
          "Terms"
        ],
        "summary": "Search the terms annotated in a field, or in every annotation field",
        "description": "Searches the terms of a field, or of every annotation field, and counts each hit in `unit`. Hits are ordered by how they match `query`, then by `count`. Each hit is counted in the population of its own field. The population is `q`. With `facetSelfExclude`, it is `q` without the conjuncts on that field. See \"Default elements\" in /llms-full.txt for the hits that `limit` keeps. Use the `termId` of a hit in a condition.",
        "operationId": "searchTerms",
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 256
                },
                {
                  "type": "null"
                }
              ],
              "description": "Annotation field, such as `disease`; omitted means every annotation field",
              "title": "Field"
            },
            "description": "Annotation field, such as `disease`; omitted means every annotation field"
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4096,
              "description": "Text to find in the label, a synonym, or the ID of a term, ignoring letter case; empty lists the most annotated terms. This is not the condition `q`",
              "default": "",
              "title": "Query"
            },
            "description": "Text to find in the label, a synonym, or the ID of a term, ignoring letter case; empty lists the most annotated terms. This is not the condition `q`"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt",
              "examples": [
                "disease:\"MONDO:0007254\"",
                "cell_line:\"CVCL:0030\" AND library_strategy:ATAC-seq",
                "organism_id:9606 AND date_published:[2015-01-01 TO 2020-12-31]",
                "disease_status:unmapped",
                "\"breast cancer\" AND NOT organism_id:10090"
              ],
              "title": "Q"
            },
            "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt"
          },
          {
            "name": "unit",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Unit",
              "description": "Counting unit. `biosample` counts distinct BioSamples, `sra-experiment` distinct SRA Experiments, and `bioproject` distinct BioProjects. See \"Counting\" in /llms-full.txt",
              "default": "biosample"
            },
            "description": "Counting unit. `biosample` counts distinct BioSamples, `sra-experiment` distinct SRA Experiments, and `bioproject` distinct BioProjects. See \"Counting\" in /llms-full.txt"
          },
          {
            "name": "facetSelfExclude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, the operation removes from `q` each top-level conjunct whose clauses are all on the dimensions of the operation, as the UI does. The top-level conjuncts are the operands of the `AND` at the root of `q`, or `q` itself if the root is not `AND`. A conjunct that has a keyword stays. The description of the operation names its dimensions. If false, the operation uses `q` as it is. See \"Self-exclusion\" in /llms-full.txt",
              "default": false,
              "title": "Facetselfexclude"
            },
            "description": "If true, the operation removes from `q` each top-level conjunct whose clauses are all on the dimensions of the operation, as the UI does. The top-level conjuncts are the operands of the `AND` at the root of `q`, or `q` itself if the root is not `AND`. A conjunct that has a keyword stays. The description of the operation names its dimensions. If false, the operation uses `q` as it is. See \"Self-exclusion\" in /llms-full.txt"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of hits to return, from 1 to 100",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of hits to return, from 1 to 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. The slug that ends `type` is one of:\n\n- `unexpected-token`: The condition does not parse, is longer than 4096 characters, or is blank where a condition is required, or a keyword has a wildcard. Parse, select, and keyword also return this slug if the condition that they return would be longer than 4096 characters. Fix the condition at the column that `detail` gives. Quote a value that has a colon or a space. See \"Grammar\" in /llms-full.txt.\n- `unknown-field`: A condition, a clause, or a parameter (`field`, `row`, `col`) names a field that the dataset does not have. Use a field that `dslFields` of `GET /api/dataset` lists.\n- `invalid-date-format`: A value of `date_published` in the form `YYYY-MM-DD` is not a calendar date, such as `2020-13-45`, or a bound of a range is not a date in that form. Write the date as `YYYY-MM-DD`, such as `2020-12-31`. See \"Fields\" in /llms-full.txt.\n- `invalid-operator-for-field`: The form of a value does not fit the field, such as a range or a date on a field other than `date_published`, a value with a wildcard, or a value of `date_published` that is neither a date in the form `YYYY-MM-DD` nor a range (`2020-1-1`, a quoted date, or `\"\"`). Use `field:value` for the field, `date_published:YYYY-MM-DD` for a date, or `date_published:[a TO b]` for a range. See \"Fields\" in /llms-full.txt.\n- `invalid-value`: A value is not allowed for the field: a status that is not a status group, an assay that the dataset does not have, a term value that is not a term ID, an organism that is not a number, a keyword without a letter or a digit, or keywords over their limits. Use the values that `detail` lists. Find a term ID with `GET /api/terms`. See \"Fields\" and \"Keywords\" in /llms-full.txt.\n- `nest-depth-exceeded`: Groups nest deeper than 5 levels, or the condition has more than 512 nodes. Flatten the groups or split the condition. See \"Limits\" in /llms-full.txt.\n- `missing-value`: A clause on a field other than `date_published` has an empty value. Give the clause a value. See \"Fields\" in /llms-full.txt.\n- `invalid-dimension`: A field exists but cannot be the dimension of the operation, or the two dimensions of a cross-tabulation are the same field. Choose from the fields that `detail` lists.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable. The slug that ends `type` is one of:\n\n- `server-busy`: The api has no free slot for the request after waiting, or too many requests wait already. Wait for the `Retry-After` seconds and send the request again.\n- `query-timeout`: The query took longer than the time limit of a request and was stopped. Narrow the condition. The same request does not finish by itself.\n- `query-too-large`: The query needs more memory or temporary disk space than a worker may use and was stopped. Narrow the condition, or ask for fewer elements.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/terms/children": {
      "get": {
        "tags": [
          "Terms"
        ],
        "summary": "List the child terms of a term annotated in a field",
        "description": "Lists the direct child terms of `termId` that have a count above 0 in the population, in `unit`, in the order of their direct counts in the whole dataset. `count` of a child includes its descendants. The population is `q`. With `facetSelfExclude`, it is `q` without the conjuncts on `field`. To expand a term of a cross-tabulation, pass its `populationQ` as `q`. See \"Aggregations\" in /llms-full.txt.",
        "operationId": "listTermChildren",
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "description": "Annotation field of the term, such as `disease`",
              "title": "Field"
            },
            "description": "Annotation field of the term, such as `disease`"
          },
          {
            "name": "termId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "description": "ID of the term, such as `MONDO:0007254`",
              "title": "Termid"
            },
            "description": "ID of the term, such as `MONDO:0007254`"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt",
              "examples": [
                "disease:\"MONDO:0007254\"",
                "cell_line:\"CVCL:0030\" AND library_strategy:ATAC-seq",
                "organism_id:9606 AND date_published:[2015-01-01 TO 2020-12-31]",
                "disease_status:unmapped",
                "\"breast cancer\" AND NOT organism_id:10090"
              ],
              "title": "Q"
            },
            "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt"
          },
          {
            "name": "unit",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Unit",
              "description": "Counting unit. `biosample` counts distinct BioSamples, `sra-experiment` distinct SRA Experiments, and `bioproject` distinct BioProjects. See \"Counting\" in /llms-full.txt",
              "default": "biosample"
            },
            "description": "Counting unit. `biosample` counts distinct BioSamples, `sra-experiment` distinct SRA Experiments, and `bioproject` distinct BioProjects. See \"Counting\" in /llms-full.txt"
          },
          {
            "name": "facetSelfExclude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, the operation removes from `q` each top-level conjunct whose clauses are all on the dimensions of the operation, as the UI does. The top-level conjuncts are the operands of the `AND` at the root of `q`, or `q` itself if the root is not `AND`. A conjunct that has a keyword stays. The description of the operation names its dimensions. If false, the operation uses `q` as it is. See \"Self-exclusion\" in /llms-full.txt",
              "default": false,
              "title": "Facetselfexclude"
            },
            "description": "If true, the operation removes from `q` each top-level conjunct whose clauses are all on the dimensions of the operation, as the UI does. The top-level conjuncts are the operands of the `AND` at the root of `q`, or `q` itself if the root is not `AND`. A conjunct that has a keyword stays. The description of the operation names its dimensions. If false, the operation uses `q` as it is. See \"Self-exclusion\" in /llms-full.txt"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermChildrenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. The slug that ends `type` is one of:\n\n- `unexpected-token`: The condition does not parse, is longer than 4096 characters, or is blank where a condition is required, or a keyword has a wildcard. Parse, select, and keyword also return this slug if the condition that they return would be longer than 4096 characters. Fix the condition at the column that `detail` gives. Quote a value that has a colon or a space. See \"Grammar\" in /llms-full.txt.\n- `unknown-field`: A condition, a clause, or a parameter (`field`, `row`, `col`) names a field that the dataset does not have. Use a field that `dslFields` of `GET /api/dataset` lists.\n- `invalid-date-format`: A value of `date_published` in the form `YYYY-MM-DD` is not a calendar date, such as `2020-13-45`, or a bound of a range is not a date in that form. Write the date as `YYYY-MM-DD`, such as `2020-12-31`. See \"Fields\" in /llms-full.txt.\n- `invalid-operator-for-field`: The form of a value does not fit the field, such as a range or a date on a field other than `date_published`, a value with a wildcard, or a value of `date_published` that is neither a date in the form `YYYY-MM-DD` nor a range (`2020-1-1`, a quoted date, or `\"\"`). Use `field:value` for the field, `date_published:YYYY-MM-DD` for a date, or `date_published:[a TO b]` for a range. See \"Fields\" in /llms-full.txt.\n- `invalid-value`: A value is not allowed for the field: a status that is not a status group, an assay that the dataset does not have, a term value that is not a term ID, an organism that is not a number, a keyword without a letter or a digit, or keywords over their limits. Use the values that `detail` lists. Find a term ID with `GET /api/terms`. See \"Fields\" and \"Keywords\" in /llms-full.txt.\n- `nest-depth-exceeded`: Groups nest deeper than 5 levels, or the condition has more than 512 nodes. Flatten the groups or split the condition. See \"Limits\" in /llms-full.txt.\n- `missing-value`: A clause on a field other than `date_published` has an empty value. Give the clause a value. See \"Fields\" in /llms-full.txt.\n- `invalid-dimension`: A field exists but cannot be the dimension of the operation, or the two dimensions of a cross-tabulation are the same field. Choose from the fields that `detail` lists.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable. The slug that ends `type` is one of:\n\n- `server-busy`: The api has no free slot for the request after waiting, or too many requests wait already. Wait for the `Retry-After` seconds and send the request again.\n- `query-timeout`: The query took longer than the time limit of a request and was stopped. Narrow the condition. The same request does not finish by itself.\n- `query-too-large`: The query needs more memory or temporary disk space than a worker may use and was stopped. Narrow the condition, or ask for fewer elements.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/terms/{termId}": {
      "get": {
        "tags": [
          "Terms"
        ],
        "summary": "Get a term with its synonyms, parents, ontology, and page",
        "description": "Returns one term of the dataset: its label, synonyms, direct parent terms, ontology, and the address of its page. `parents` is empty for a root of an ontology and for a term of an ontology without hierarchy. See \"Term hierarchy\" in /llms-full.txt.",
        "operationId": "getTerm",
        "parameters": [
          {
            "name": "termId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "description": "ID of the term, such as `UBERON:0000955`",
              "title": "Termid"
            },
            "description": "ID of the term, such as `UBERON:0000955`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TermsResponse": {
        "properties": {
          "datasetVersion": {
            "$ref": "#/components/schemas/DatasetVersionRef"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "The searched field, or null when every annotation field was searched"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "populationQ": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Populationq",
            "description": "The condition the counts were computed from, or null when every annotation field was searched"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/TermHit"
            },
            "type": "array",
            "title": "Terms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "datasetVersion",
          "field",
          "query",
          "populationQ",
          "unit",
          "terms"
        ],
        "title": "TermsResponse"
      },
...
      "Unit": {
        "type": "string",
        "enum": [
          "biosample",
          "sra-experiment",
          "bioproject"
        ]
      }
    }
  },
...
}